Navigation bar was implemented.

+++ MyMoviesApplication/Forms/FormActors.cs
@@ -17,7 +17,7 @@ namespace MyMoviesApplication.Forms
     public partial class FormActors : Form
     {        
         private readonly Classes.Actors actors = new Classes.Actors();
-
+       
         //*********************************************************************************************************
         //Private methods
         private void ClearControls()
@@ -37,12 +37,14 @@ namespace MyMoviesApplication.Forms
 
         private void LinkDataGridViewToFields()
         {
+            toolStripLabelRowsIdentification.Text = "{ " + Convert.ToInt32(dataGridViewActors.CurrentRow.Index + 1) + " } of " + dataGridViewActors.Rows.Count;
+
             textBoxActorId.Text = dataGridViewActors.Rows[dataGridViewActors.CurrentRow.Index].Cells[0].Value.ToString();
             textBoxActor.Text = dataGridViewActors.Rows[dataGridViewActors.CurrentRow.Index].Cells[1].Value.ToString();
             textBoxCredits.Text = dataGridViewActors.Rows[dataGridViewActors.CurrentRow.Index].Cells[2].Value.ToString();
             textBoxLinkImdb.Text = dataGridViewActors.Rows[dataGridViewActors.CurrentRow.Index].Cells[3].Value.ToString();
             dateTimePickerRegisterDate.Value = Convert.ToDateTime(dataGridViewActors.Rows[dataGridViewActors.CurrentRow.Index].Cells[4].Value.ToString());
-            dateTimePickerLastUpdate.Value = Convert.ToDateTime(dataGridViewActors.Rows[dataGridViewActors.CurrentRow.Index].Cells[5].Value.ToString());
+            dateTimePickerLastUpdate.Value = Convert.ToDateTime(dataGridViewActors.Rows[dataGridViewActors.CurrentRow.Index].Cells[5].Value.ToString());                       
         }
 
         //**************************************************************************************************************
@@ -51,7 +53,7 @@ namespace MyMoviesApplication.Forms
         public FormActors()
         {
             InitializeComponent();
-            LoadActorsTable();
+            LoadActorsTable();           
         }
                        
         private void DataGridViewActors_SelectionChanged(object sender, EventArgs e)
@@ -62,24 +64,49 @@ namespace MyMoviesApplication.Forms
         private void toolStripButtonAdd_Click(object sender, EventArgs e)
         {
             ClearControls();
+            toolStripButtonAdd.Enabled = false;
+            toolStripButtonUpdate.Enabled = false;
+            toolStripButtonDelete.Enabled = false;
+            toolStripButtonSave.Enabled = true;
+            dateTimePickerRegisterDate.Value = DateTime.Now;
+            dateTimePickerLastUpdate.Value = DateTime.Now;
         }
 
         private void toolStripButtonSave_Click(object sender, EventArgs e)
         {
             actors.InsertRow(textBoxActor.Text, Convert.ToInt32(textBoxCredits.Text), textBoxLinkImdb.Text, dateTimePickerRegisterDate.Value, dateTimePickerRegisterDate.Value);
-            LoadActorsTable();            
+            LoadActorsTable();
+            toolStripLabelRowsIdentification.Text = "{ " + Convert.ToInt32(dataGridViewActors.CurrentRow.Index + 1) + " } of " + dataGridViewActors.Rows.Count;
+            toolStripButtonAdd.Enabled = true;
+            toolStripButtonUpdate.Enabled = true;
+            toolStripButtonDelete.Enabled = true;
+            toolStripButtonSave.Enabled = false;
         }
 
-        private void toolStripButtonRow_Click(object sender, EventArgs e)
+        private void toolStripButtonDelete_Click(object sender, EventArgs e)
         {
-            actors.DeleteRow(Convert.ToInt32(textBoxActorId.Text));
-            LoadActorsTable();
+            var question = MessageBox.Show("Are you sure that you want to delete this row?", "Question", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            
+            if(question == DialogResult.Yes)
+            {
+                actors.DeleteRow(Convert.ToInt32(textBoxActorId.Text));
+                LoadActorsTable();
+                toolStripLabelRowsIdentification.Text = "{ " + Convert.ToInt32(dataGridViewActors.CurrentRow.Index + 1) + 
+                                                        " } of " + dataGridViewActors.Rows.Count;
+            }
+           
         }
 
         private void toolStripButtonUpdate_Click(object sender, EventArgs e)
         {
             actors.Update(Convert.ToInt32(textBoxActorId.Text), textBoxActor.Text, Convert.ToInt32(textBoxCredits.Text), textBoxLinkImdb.Text, dateTimePickerRegisterDate.Value, dateTimePickerLastUpdate.Value);
             LoadActorsTable();
+            toolStripLabelRowsIdentification.Text = "{ " + Convert.ToInt32(dataGridViewActors.CurrentRow.Index + 1) + " } of " + dataGridViewActors.Rows.Count;
+        }
+               
+        private void FormActors_Shown(object sender, EventArgs e)
+        {
+            toolStripLabelRowsIdentification.Text = "{ " + Convert.ToInt32(dataGridViewActors.CurrentRow.Index + 1) + " } of " + dataGridViewActors.Rows.Count;
         }
     }
 }
